@@ -7,7 +7,7 @@ import {HeaderComponent} from './base/header/header.component';
 import {FooterComponent} from './base/footer/footer.component';
 import {HomeComponent} from './components/home/home.component';
 import {ResultComponent} from './components/result/result.component';
-import {GaugeModule, NgxChartsModule} from '@swimlane/ngx-charts';
+import {NgxChartsModule} from '@swimlane/ngx-charts';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {RuleSetComponent} from './components/result/rules-set/rule-set.component';
 import {CollapseModule} from 'ngx-bootstrap';
@@ -15,8 +15,10 @@ import {SubCategoryComponent} from './components/result/sub-category/sub-categor
 import {FormsModule} from '@angular/forms';
 import {IssuesComponent} from './components/result/issues/issues.component';
 import {PieGridComponent} from './components/pie-grid/pie-grid.component';
-import { HistoryComponent } from './components/history/history.component';
-import {HttpClientModule} from '@angular/common/http';
+import {HistoryComponent} from './components/history/history.component';
+import {HTTP_INTERCEPTORS, HttpClientModule} from '@angular/common/http';
+import {AppService} from './service/app.service';
+import {RequestInterceptor} from './interceptors/request-interceptor.service';
 
 @NgModule({
   declarations: [
@@ -40,7 +42,7 @@ import {HttpClientModule} from '@angular/common/http';
     CollapseModule,
     HttpClientModule
   ],
-  providers: [],
+  providers: [ AppService, { provide: HTTP_INTERCEPTORS, useClass: RequestInterceptor, multi: true } ],
   bootstrap: [AppComponent]
 })
 export class AppModule {
